# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Intune Policy Auditor API",
    "description": "Read-only, deterministic analysis of exported Intune policies.",
    "version": "3.0.0-dev"
  },
  "paths": {
...
    "/api/v1/knowledge-packs/validate": {
      "post": {
        "tags": [
          "Knowledge Packs"
        ],
        "summary": "Validate Knowledge Pack",
        "operationId": "validate_knowledge_pack_api_v1_knowledge_packs_validate_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_validate_knowledge_pack_api_v1_knowledge_packs_validate_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope_PackValidationReport_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge-packs/import": {
      "post": {
        "tags": [
          "Knowledge Packs"
        ],
        "summary": "Import Knowledge Pack",
        "operationId": "import_knowledge_pack_api_v1_knowledge_packs_import_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_knowledge_pack_api_v1_knowledge_packs_import_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope_KnowledgePackSummary_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge-packs/{pack_id}/activate": {
      "post": {
        "tags": [
          "Knowledge Packs"
        ],
        "summary": "Activate Pack",
        "operationId": "activate_pack_api_v1_knowledge_packs__pack_id__activate_post",
        "parameters": [
          {
            "name": "pack_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pack Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope_KnowledgePackSummary_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge-packs/{pack_id}/deactivate": {
      "post": {
        "tags": [
          "Knowledge Packs"
        ],
        "summary": "Deactivate Pack",
        "operationId": "deactivate_pack_api_v1_knowledge_packs__pack_id__deactivate_post",
        "parameters": [
          {
            "name": "pack_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pack Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope_KnowledgePackSummary_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge-packs/{pack_id}": {
      "delete": {
        "tags": [
          "Knowledge Packs"
        ],
        "summary": "Remove Pack",
        "operationId": "remove_pack_api_v1_knowledge_packs__pack_id__delete",
        "parameters": [
          {
            "name": "pack_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pack Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope_dict_str__bool__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge-packs/{pack_id}/provenance": {
      "get": {
        "tags": [
          "Knowledge Packs"
        ],
        "summary": "Pack Provenance",
        "operationId": "pack_provenance_api_v1_knowledge_packs__pack_id__provenance_get",
        "parameters": [
          {
            "name": "pack_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pack Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope_KnowledgePackManifest_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge-packs/{pack_id}/validation-report": {
      "get": {
        "tags": [
          "Knowledge Packs"
        ],
        "summary": "Export Pack Validation Report",
        "operationId": "export_pack_validation_report_api_v1_knowledge_packs__pack_id__validation_report_get",
        "parameters": [
          {
            "name": "pack_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pack Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tenant/device-code/complete": {
      "post": {
        "tags": [
          "Tenant Mode"
        ],
        "summary": "Complete Device Code",
        "operationId": "complete_device_code_api_v1_tenant_device_code_complete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteDeviceCode"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope_AuthenticationResult_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tenant/retrieve/{operation}": {
      "get": {
        "tags": [
          "Tenant Mode"
        ],
        "summary": "Retrieve",
        "operationId": "retrieve_api_v1_tenant_retrieve__operation__get",
        "parameters": [
          {
            "name": "operation",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Operation"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope_GraphCollectionResult_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthenticationResult": {
        "properties": {
          "authenticated": {
            "type": "boolean",
            "title": "Authenticated"
          },
          "granted_scopes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Granted Scopes"
          },
          "account_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Name"
          }
        },
        "type": "object",
        "required": [
          "authenticated",
          "granted_scopes"
        ],
        "title": "AuthenticationResult"
      },
...
      "Body_import_knowledge_pack_api_v1_knowledge_packs_import_post": {
        "properties": {
          "pack": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Pack",
            "description": "Validated knowledge-pack ZIP"
          }
        },
        "type": "object",
        "required": [
          "pack"
        ],
        "title": "Body_import_knowledge_pack_api_v1_knowledge_packs_import_post"
      },
...
      "Body_validate_knowledge_pack_api_v1_knowledge_packs_validate_post": {
        "properties": {
          "pack": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Pack",
            "description": "ZIP containing one manifest.json and settings JSON"
          }
        },
        "type": "object",
        "required": [
          "pack"
        ],
        "title": "Body_validate_knowledge_pack_api_v1_knowledge_packs_validate_post"
      },
      "CompleteDeviceCode": {
        "properties": {
          "flow_id": {
            "type": "string",
            "title": "Flow Id"
          }
        },
        "type": "object",
        "required": [
          "flow_id"
        ],
        "title": "CompleteDeviceCode"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KnowledgePackManifest": {
        "properties": {
          "schema_version": {
            "type": "string",
            "title": "Schema Version"
          },
          "pack_id": {
            "type": "string",
            "pattern": "^[a-z0-9][a-z0-9._-]+$",
            "title": "Pack Id"
          },
          "vendor": {
            "type": "string",
            "minLength": 1,
            "title": "Vendor"
          },
          "product": {
            "type": "string",
            "minLength": 1,
            "title": "Product"
          },
          "baseline_name": {
            "type": "string",
            "minLength": 1,
            "title": "Baseline Name"
          },
          "baseline_version": {
            "type": "string",
            "minLength": 1,
            "title": "Baseline Version"
          },
          "platform": {
            "$ref": "#/components/schemas/Platform"
          },
          "source_type": {
            "$ref": "#/components/schemas/EvidenceType"
          },
          "source_title": {
            "type": "string",
            "minLength": 1,
            "title": "Source Title"
          },
          "source_domain": {
            "type": "string",
            "minLength": 1,
            "title": "Source Domain"
          },
          "source_reference": {
            "type": "string",
            "minLength": 1,
            "title": "Source Reference"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "verified_at": {
            "type": "string",
            "format": "date",
            "title": "Verified At"
          },
          "imported_at": {
            "type": "string",
            "format": "date-time",
            "title": "Imported At"
          },
          "data_sha256": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$",
            "title": "Data Sha256"
          },
          "schema_sha256": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$",
            "title": "Schema Sha256"
          },
          "application_version": {
            "type": "string",
            "title": "Application Version"
          },
          "status": {
            "$ref": "#/components/schemas/KnowledgePackStatus"
          },
          "setting_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Setting Count"
          },
          "exact_value_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Exact Value Count"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          },
          "settings_file": {
            "type": "string",
            "pattern": "^[^/\\\\]+\\.json$",
            "title": "Settings File"
          },
          "provenance": {
            "$ref": "#/components/schemas/KnowledgeProvenance"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "schema_version",
          "pack_id",
          "vendor",
          "product",
          "baseline_name",
          "baseline_version",
          "platform",
          "source_type",
          "source_title",
          "source_domain",
          "source_reference",
          "verified_at",
          "imported_at",
          "data_sha256",
          "schema_sha256",
          "application_version",
          "status",
          "setting_count",
          "exact_value_count",
          "settings_file"
        ],
        "title": "KnowledgePackManifest"
      },
...
      "KnowledgePackSummary": {
        "properties": {
          "pack_id": {
            "type": "string",
            "title": "Pack Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "vendor": {
            "type": "string",
            "title": "Vendor"
          },
          "product": {
            "type": "string",
            "title": "Product"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "platform": {
            "type": "string",
            "title": "Platform"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "verification_date": {
            "type": "string",
            "format": "date",
            "title": "Verification Date"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "data_sha256": {
            "type": "string",
            "title": "Data Sha256"
          },
          "setting_count": {
            "type": "integer",
            "title": "Setting Count"
          },
          "exact_value_count": {
            "type": "integer",
            "title": "Exact Value Count"
          },
          "age_days": {
            "type": "integer",
            "title": "Age Days"
          },
          "stale": {
            "type": "boolean",
            "title": "Stale"
          },
          "validation_state": {
            "type": "string",
            "title": "Validation State"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "synthetic": {
            "type": "boolean",
            "title": "Synthetic"
          }
        },
        "type": "object",
        "required": [
          "pack_id",
          "name",
          "vendor",
          "product",
          "version",
          "platform",
          "status",
          "verification_date",
          "source",
          "data_sha256",
          "setting_count",
          "exact_value_count",
          "age_days",
          "stale",
          "validation_state",
          "active",
          "synthetic"
        ],
        "title": "KnowledgePackSummary"
      },
...
      "PackValidationIssue": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "path",
          "message"
        ],
        "title": "PackValidationIssue"
      },
      "PackValidationReport": {
        "properties": {
          "valid": {
            "type": "boolean",
            "title": "Valid"
          },
          "pack_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pack Id"
          },
          "checked_at": {
            "type": "string",
            "format": "date-time",
            "title": "Checked At"
          },
          "calculated_data_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calculated Data Sha256"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/PackValidationIssue"
            },
            "type": "array",
            "title": "Issues"
          }
        },
        "type": "object",
        "required": [
          "valid"
        ],
        "title": "PackValidationReport"
      },
...
      "ResponseEnvelope_AuthenticationResult_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AuthenticationResult"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "ResponseEnvelope[AuthenticationResult]"
      },
...
      "ResponseEnvelope_GraphCollectionResult_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/GraphCollectionResult"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "ResponseEnvelope[GraphCollectionResult]"
      },
...
      "ResponseEnvelope_KnowledgePackManifest_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/KnowledgePackManifest"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "ResponseEnvelope[KnowledgePackManifest]"
      },
      "ResponseEnvelope_KnowledgePackSummary_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/KnowledgePackSummary"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "ResponseEnvelope[KnowledgePackSummary]"
      },
...
      "ResponseEnvelope_PackValidationReport_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PackValidationReport"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "ResponseEnvelope[PackValidationReport]"
      },
...
      "ResponseEnvelope_dict_str__bool__": {
        "properties": {
          "data": {
            "additionalProperties": {
              "type": "boolean"
            },
            "type": "object",
            "title": "Data"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "ResponseEnvelope[dict[str, bool]]"
      },
...
      "ResponseMeta": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          }
        },
        "type": "object",
        "required": [
          "request_id"
        ],
        "title": "ResponseMeta"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}